% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_ae_be.R
\name{americanize_exact}
\alias{americanize_exact}
\title{Convert British English to Amrican English}
\usage{
americanize_exact(string, stand.table = NULL, nthreads = 1)
}
\arguments{
\item{string}{The input string (the string must be lemmatized)}

\item{stand.table}{A table with a pair of words in British English and American English \cr
Note: If no table is specified the default table will be used \cr}

\item{nthreads}{}
}
\value{
A string with only American English spelling
}
\description{
Convert British English to Amrican English
}
