% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations.R
\name{americanize_fast}
\alias{americanize_fast}
\title{Convert British English to Amrican English (Fast Version)}
\usage{
americanize_fast(string, replace.table = NULL)
}
\arguments{
\item{string}{The input string}

\item{replace.table}{A table with British English and American English spelling}
}
\value{
A string with only American English spelling
}
\description{
Takes the input string and replaces British spelling with American spelling. In this
fast implementation of the function the exact punctuation (e.g. blanks) is not preserved
}
\examples{
americanize_fast(c("capitalised asset", "profit centre"))

}
