% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_bind_rows.R
\name{bind_df_list}
\alias{bind_df_list}
\title{Bind dataframes contained in list with document id}
\usage{
bind_df_list(list, names, id)
}
\arguments{
\item{list}{List that contains dataframes}

\item{names}{The names of the dataframes}

\item{id}{Name of the id column}
}
\value{
A dataframe
}
\description{
Bind dataframes contained in list with document id
}
