% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_operations.R
\name{check_repl_list}
\alias{check_repl_list}
\title{Check Input Tables for Word Replacement}
\usage{
check_repl_list(replace_table)
}
\arguments{
\item{replace_table}{A Replacement Table (needs 2 columns: "word" and "replace")}
}
\value{
Errors
}
\description{
Check Input Tables for Word Replacement
}
