% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_functions.R
\name{con_dedup}
\alias{con_dedup}
\title{Deduplicate Concepts}
\usage{
con_dedup(df, col.term, col.conid)
}
\arguments{
\item{df}{A dataframe containing terms and corresponding concept ids}

\item{col.term}{The name of the column containing the terms as character string}

\item{col.conid}{The name of the column containing the concept ids as character string}
}
\value{
A dataframe with deduplicated concepts
}
\description{
This function checks if two concepts in the dataframe contain exactly the same terms
and keeps only one of the concepts
}
\examples{
library(tpfuns)
test <- tibble::tibble(
con_id = c( 1,   1,   1,   2,   2,   3,   3,   4,   4 ,  5,   5,   5,   6,   6 ),
term   = c("a", "c", "b", "a", "b", "d", "e", "e", "f", "e", "f", "g", "b", "a"))

con_dedup(test, "term", "con_id")
}
