% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_functions.R
\name{con_disambig}
\alias{con_disambig}
\title{Disambiguate Concepts}
\usage{
con_disambig(df, col.term, col.conid)
}
\arguments{
\item{df}{A dataframe containing terms and corresponding concept ids}

\item{col.term}{The name of the column containing the terms as character string}

\item{col.conid}{The name of the column containing the concept ids as character string}
}
\value{
A dataframe with disambiguated concepts
}
\description{
#' @description
This function checks two things:\cr
1.: Whether a concept is a subset of another concept. In this case the short concept is
deleted while the longer concept is preserved\cr
2.: Whether concepts share terms without being subsets of another. In this case all
concepts are deleted\cr#'
IMPORTANT NOTE: Concepts should be deduplicated before disambiguated!!
}
\examples{
library(tpfuns)
test <- tibble::tibble(
con_id = c( 1,   1,   1,   2,   2,   3,   3,   4,   4 ,  5,   5,   5,   6,   6 ),
term   = c("a", "c", "b", "a", "b", "d", "e", "e", "f", "e", "f", "g", "b", "a"))

con_disambig(test, "term", "con_id")
}
