% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_functions.R
\name{con_merge}
\alias{con_merge}
\title{Merge Concepts}
\usage{
con_merge(df, col.term, col.conid)
}
\arguments{
\item{df}{A dataframe containing terms and corresponding concept ids}

\item{col.term}{The name of the column containing the terms as character string}

\item{col.conid}{The name of the column containing the concept ids as character string}
}
\value{
A dataframe with merged concepts (reassigned concept ids)
}
\description{
this function reassigns concept ids if different concepts contain same terms
}
\examples{
library(tpfuns)
test <- tibble::tibble(
con_id = c( 1,   1,   1,   2,   2,   3,   3,   4,   4 ,  5,   5,   5,   6,   6 ),
term   = c("a", "c", "b", "a", "b", "d", "e", "e", "f", "e", "f", "g", "b", "a"))

con_merge(test, "term", "con_id")
}
