% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_operations.R
\name{cop_rem_le}
\alias{cop_rem_le}
\title{Remove Legal Entities from String}
\usage{
cop_rem_le(string, le.return = FALSE)
}
\arguments{
\item{string}{A chracter string}
}
\value{
A character String without legal entities
}
\description{
This function is part of the 'company operations' (cop) function set\cr
This function removes legal entities from company names
}
\examples{
tpfuns::cop_rem_le("basf se")
}
