% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_operations.R
\name{cop_repl_words}
\alias{cop_repl_words}
\title{Standardize Commonly Misspelled Words in Strings}
\usage{
cop_repl_words(string, replace_table = NULL)
}
\arguments{
\item{string}{A charcter string}

\item{replace_table}{A Replacement Table (needs 2 columns: "word" and "replace")\cr
If NULL the default table will be used}
}
\value{
A character string with standardized words
}
\description{
This function is part of the 'company operations' (cop) function set\cr
This function replaces and standardizes commonly misspelled words in company names
}
\examples{
cop_repl_words("1st")
}
