% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_operations.R
\name{cop_stand_le}
\alias{cop_stand_le}
\title{Remove Legal Entities from String}
\usage{
cop_stand_le(string, le.op = c("remove", "stand"),
  table.return = FALSE)
}
\arguments{
\item{string}{A chracter string}

\item{le.op}{Legal Entity Operation: Should the legal entity be removed or standardized?}

\item{table.return}{Should the whole table with legal entity information be returned?}
}
\value{
Either a string in the case of le.op = "remove" or le.op = "stand", or a dataframe in
case le.op = c("remove", "stand") or table.return = TRUE
}
\description{
This function is part of the 'company operations' (cop) function set\cr
This function standardizes or removes legal entities from character strings with company
name
}
\examples{
companies <- c("basf co ltd", "basf company ltd", "basf ag", "basf se", "basf aktiengesellschaft")
tpfuns::cop_stand_le(companies, le.op = "remove")
tpfuns::cop_stand_le(companies, le.op = "stand")
tpfuns::cop_stand_le(companies)
tpfuns::cop_stand_le(companies,  table.return = TRUE)

}
