% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_functions.R
\name{copm_match}
\alias{copm_match}
\title{Match company names against Databses}
\usage{
copm_match(names, match.table = NULL, match.type = c("full", "sub",
  "approx"), nthreads = 1)
}
\arguments{
\item{names}{A charcter string of company names}

\item{match.table}{A table prepared by copm_prep_names}

\item{match.type}{Any combination of c("full", "sub", "approx")}
}
\value{
A list with matches
}
\description{
Match company names against Databses
}
\examples{
library(tpfuns)
table <- tibble::tibble(id = c(1,2,3), firm = c("BASF GMBH", "BASF AG", "BASF SE (GER)"))
match.table <- copm_prep_db(table, "firm", "id")
names <- c("BASF GMBH", "BASF AG", "Siemens AG")
copm_match(names, match.table, "full")
}
