% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_match.R
\name{copm_prep_db}
\alias{copm_prep_db}
\title{Preparation of Database names for Company Matching Procedure}
\usage{
copm_prep_db(match.table, col.name = NULL, col.ident = NULL,
  adj.regex = "default")
}
\arguments{
\item{match.table}{A table with company names and identifiers}

\item{col.name}{A chracter string of the name column}

\item{col.ident}{A character string of the identifier columns}

\item{adj.regex}{Adjustemtn REgex}
}
\value{
A dataframe suitable for the matching algorithm
}
\description{
Preparation of Database names for Company Matching Procedure
}
