% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_functions.R
\name{copm_prep_names}
\alias{copm_prep_names}
\title{Name Preparation for Company Matching Procedure}
\usage{
copm_prep_names(names,
  adj.regex = "\\\\s?holdings?\\\\s?|\\\\s?groups?\\\\s?|\\\\s?units?$")
}
\arguments{
\item{names}{A character string of company names}

\item{adj.regex}{Adjustment Regex}
}
\value{
A dataframe with different 3 columns containing different name varieties.\cr
Col 1 ('n0'):    Original Name of the company.\cr
Col 2 ('n1'):    Adjusted Name of the company.\cr
Col 3 ('ntype'): Adjustemt type.
}
\description{
This function takes a character string and prepares a deduplicated list with maximum 24
different variations of the name. The different name varieties are stored in the 'ntype'
}
\examples{
library(tpfuns)
copm_prep_names(names = c("BASF GMBH", "BASF AG", "BASF SE (GER)"))
}
