% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{cos_string_pair}
\alias{cos_string_pair}
\title{Cosine Similarity between two Strings}
\usage{
cos_string_pair(a, b, rem.stop = FALSE)
}
\arguments{
\item{a}{A String (or a vector of strings)}

\item{b}{A String (or a vector of strings)}

\item{rem.stop}{Shall stopwords be removed?}
}
\value{
The cosine similarity (either a single number or a vector of numbers)
}
\description{
Cosine Similarity between two Strings
}
