% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_pdf.R
\name{decrypt_pdf}
\alias{decrypt_pdf}
\title{Decrypt pdfs}
\usage{
decrypt_pdf(file.in, file.out)
}
\arguments{
\item{path.in}{complete path to the encrypted pdf file
(must be different than path.out)}

\item{path.out}{complete path where the decrypted file should be written
(must be different than path.in)}
}
\value{
A pdf file (decrypted)
}
\description{
A wrapper around the qpdf command line tool.
First, the qpdf tool must be downloaded (https://sourceforge.net/projects/qpdf/files/qpdf/).
Second, zipped file must be extracted to any folder.
Third, environment variables must be updated (...\qpdf-x.x.x\bin).
Program is tested on qpdf version 8.1.0.
}
