% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_statistics.R
\name{docs_stats}
\alias{docs_stats}
\title{Staistics for text documents}
\usage{
docs_stats(path)
}
\arguments{
\item{path}{The path to the textfiles (get with list.files())}
}
\value{
A Dataframe
}
\description{
This function counts the number of pages (n_page), the number of words (n_words) and
the number of characters (n_char). Based on this the ration words per page (n_word_page)
and characters per words (n_char_word) is calculated
}
