% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_regex.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape Punctuation for Regex}
\usage{
escape_regex(string)
}
\arguments{
\item{string}{A String}
}
\value{
A String
}
\description{
Escape Punctuation for Regex
}
