% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fop_functions.R
\name{fop_freq_adjust}
\alias{fop_freq_adjust}
\title{Adjust frequency for substring occurances}
\usage{
fop_freq_adjust(df, group.ind, term.col, count.col)
}
\arguments{
\item{df}{The dataframe (output from the unadjusted frequency count)}

\item{group.ind}{The groups in which each adjustemt should take place (e.g. per document and sentence).
This must be an integer vector with the same length as the input table}

\item{term.col}{The column that contains the terms}

\item{count.col}{The column that contains teh unadjusted frequencies for the term column}
}
\value{
A numeric vector (adjusted frequencies)
}
\description{
Adjust frequency for substring occurances
}
\examples{
library(tpfuns)
table <- tibble::tibble(
doc_id = c(1, 1, 2, 2),
term = c("property plant and equipment", "property", "fixed asset", "asset"),
count = c(10, 12, 20, 51)
)

table$adjust <- fop_freq_adjust(table, dplyr::group_indices(table, doc_id), "term", "count")
table
}
