% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fop_functions.R
\name{fop_string_mat_count}
\alias{fop_string_mat_count}
\title{Occurance Count of Substrings}
\usage{
fop_string_mat_count(string, names = FALSE, tri.lower = TRUE)
}
\arguments{
\item{string}{A character string with elements that are substrings of other elements}

\item{names}{Shall the matrix have row/colmun names?}

\item{tri.lower}{Shall only the lower triangle have non-zero values?}
}
\value{
A squared matrix with row/column number equal the length of the input string
}
\description{
Occurance Count of Substrings
}
\examples{
library(tpfuns)
a <- c("property plant and equipment", "plant and equipment", "plant")
fop_string_mat_count(a)
fop_string_mat_count(a, TRUE)
fop_string_mat_count(a, TRUE, FALSE)

}
