% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fop_functions.R
\name{fop_tok}
\alias{fop_tok}
\title{TBD}
\usage{
fop_tok(input = NULL, ident = NULL, tokenize = c("page", "paragraph",
  "sentence", "word"), add.abbr = NULL)
}
\arguments{
\item{input}{TBD}

\item{ident}{TBD}

\item{tokenize}{TBD}

\item{add.abbr}{TBD}
}
\value{
TBD
}
\description{
TBD
}
