% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_count.R
\name{get_freq}
\alias{get_freq}
\title{Frequency Count for Terms in Text}
\usage{
get_freq(docs, terms, stand = FALSE)
}
\arguments{
\item{docs}{A vector of paths to document text files (possibly retrieved by list.files())}

\item{terms}{Either a vector of terms (if stand == FALSE) or a dataframe with the following columns: \cr
term: unstandardized terms\cr
term_stand: standardized terms\cr
term_stand_space: standardized terms without blanks (will be used for the frequency count)\cr
ngram: numbe rof words for the standardized term (term_stand)\cr
nchar: number of characters for the standardized term (term_stand)\cr

If a vector with terms is provided the function prepare_term_list will be called first.
This function returns a dataframe with the columns shown above}

\item{stand}{Are the terms already standardized (see description of 'terms' above)? \cr
If TRUE a dataframe must be provided \cr
If FALSE terms will be standardized}
}
\value{
A dataframe with adjusted and unadjusted frequency counts
}
\description{
This function takes a term list (either standardized or unstandardized) and counts the
occurances of terms in text files
}
