% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdf_info.R
\name{get_pdf_info}
\alias{get_pdf_info}
\title{Extracts metadata from pdf}
\usage{
get_pdf_info(file)
}
\arguments{
\item{file}{complete path to the pdf file}
}
\value{
a dataframe with 4 columns (readable (logical), n_page (integer),
encrypted (logical), locked (logical))
}
\description{
Extracts metadata from pdf
}
