% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemmatization.R
\name{lem_term_tg}
\alias{lem_term_tg}
\title{Wrapper around the TreeTagger Lemmatizer}
\usage{
lem_term_tg(terms, lan = c("en", "de", "es", "it", "fr"), pos = FALSE)
}
\arguments{
\item{terms}{A character string of terms to be tagged and lemmatized}

\item{lan}{The language of the terms}

\item{pos}{Shall the part of speech tags be returned? Only reasonable for small terms not for
whole texts!}
}
\value{
A dataframe with 3 columns:\cr
term:  The term to be lemmatized\cr
lemma: The lemmatized term\cr
pos:   The part of speech taggings for the tokens within the term\cr
Note:  There is some standardization of punctuation performed, so the 'term' column
does not necessarily hase the same punctuation as the original term. Nonetheless, the
order of the terms is preserved!
}
\description{
This function uses the TreeTagger Lemmatizer and is especially designed for term list
(i.e. seperate rows of terms are lemmatized and the POS tagging is returned)
}
\examples{
lem_term_tg(c("assets and liabilities", "something else, lots of"))
lem_term_tg(c("assets and liabilities", "something else, lots of"), pos = TRUE)
}
