% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations.R
\name{lemmatize_fast}
\alias{lemmatize_fast}
\title{Lemmatization of english words}
\usage{
lemmatize_fast(string)
}
\arguments{
\item{string}{The input string}
}
\value{
A character vector with lemmatized words
}
\description{
Lemmatization of english words
}
\examples{
lemmatize_fast(c("assets", "depreciated"))
}
