% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_tbl.R
\name{list_files_tbl}
\alias{list_files_tbl}
\title{List files in a dataframe with id column}
\usage{
list_files_tbl(path, file.type, id)
}
\arguments{
\item{path}{path to the files}

\item{file.type}{file ending (e.g. txt, pdf, csv, ...)}

\item{id}{column name containing the file name}
}
\value{
a dataframe with two columns
}
\description{
A helper function that is build around list.files (base)
}
