% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{lower_ws}
\alias{lower_ws}
\title{Trim and replace whitespaces}
\usage{
lower_ws(string)
}
\arguments{
\item{string}{A character of character vector}
}
\description{
A simple wrapper around some stringi function \cr
- Replaces multiple blanks with a single blank \cr
- Trims both sides of the string \cr
- Convert all characters to lower case letters
}
