% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_comp.R
\name{match_comp}
\alias{match_comp}
\title{Match company names}
\usage{
match_comp(a, b, approx = FALSE)
}
\arguments{
\item{a}{source string}

\item{b}{target string}

\item{approx}{Should approximate matching be included (takes significantly longer than exact matches)}
}
\value{
A dataframe containing matches and similarities
}
\description{
Match company names
}
\examples{
# Company name vectors for matching
source <- c("BASF", "SIEMENS") # source vector
target <- c("BASF AG", "BASFSE", "SIEMENS AG", "BASF") # target vector

# Company matching !without! approximate matches
match_comp(a = source, b = target)

# Company matching !with! approximate matches
match_comp(a = source, b = target, approx = TRUE)
}
