% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mes_functions.R
\name{mes_cos}
\alias{mes_cos}
\title{Cosine Similarity}
\usage{
mes_cos(freq.table, obj = c("normal", "matrix"),
  normalize = c("tf-idf", "none"))
}
\arguments{
\item{freq.table}{A Table}

\item{obj}{Shall a matrix of dataframe be returned}

\item{normalize}{none or tf-idf}
}
\value{
Table or Matrix
}
\description{
Cosine Similarity
}
\examples{
library(tpfuns)
freq.table <- tibble::tibble(
doc_id = c(1,1,1,1,1,2,2,2,2,3,3,3,3),
term   = c("a", "b", "c", "d", "e", "a", "b", "f", "g", "a", "n", "e", "g"),
freq   = c(21,12,58,32,14,21,14,66,14,12,85,100,12))
mes_cos(freq.table)
}
