% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mes_functions.R
\name{mes_uni_mat}
\alias{mes_uni_mat}
\title{Extract Unique Pairs from Matrix}
\usage{
mes_uni_mat(mat)
}
\arguments{
\item{mat}{A named matrix}
}
\value{
A dataframe with unique pairs
}
\description{
Extract Unique Pairs from Matrix
}
