% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{oth_bu_name}
\alias{oth_bu_name}
\title{BACKUP NAME creation}
\usage{
oth_bu_name(dir, file)
}
\arguments{
\item{dir}{The folder where the file should be saved}

\item{file}{The file name with file ending}
}
\value{
A character string
}
\description{
BACKUP NAME creation
}
\examples{
tpfuns::oth_bu_name("test_folder", "file.csv")
}
