% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{par_start}
\alias{par_start}
\title{Make Cluster and Progress Bar for PBAPPLY functions}
\usage{
par_start(nthread = 1)
}
\arguments{
\item{nthread}{Number of cores to be used}
}
\value{
Workers in the object 'cl'
}
\description{
Make Cluster and Progress Bar for PBAPPLY functions
}
