% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_csv.R
\name{pdf_to_csv}
\alias{pdf_to_csv}
\title{Converts a PDF to a tokenized csv table}
\usage{
pdf_to_csv(file, outdir)
}
\arguments{
\item{outdir}{The folder where the txt file should be saved (If NULL a new folder
called 'txt' will be created in the same directory where the pdf file is located)}

\item{A}{PDF file}
}
\value{
A csv file
}
\description{
Converts a PDF to a tokenized csv table
}
