% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_txt.R
\name{pdf_to_txt}
\alias{pdf_to_txt}
\title{Converts a PDF to text file}
\usage{
pdf_to_txt(file, outdir = NULL)
}
\arguments{
\item{file}{A PDF file}

\item{outdir}{The folder where the txt file should be saved (If NULL a new folder
called 'txt' will be created in the same directory where the pdf file is located)}
}
\value{
A text file (the text file will have the same name as the pdf file)
}
\description{
Converts a PDF to text file
}
