% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_functions.R
\name{pop_decrypt_pdf}
\alias{pop_decrypt_pdf}
\title{Decrypt pdfs}
\usage{
pop_decrypt_pdf(file.in, file.out)
}
\arguments{
\item{path.in}{complete path to the encrypted pdf file
(must be different than path.out)}

\item{path.out}{complete path where the decrypted file should be written
(must be different than path.in)}
}
\value{
A pdf file (decrypted)
}
\description{
A wrapper around the qpdf command line tool.
- First:  The qpdf tool must be downloaded (https://sourceforge.net/projects/qpdf/files/qpdf/).\cr
- Second: The zipped file must be extracted.\cr
- Third, environment variables must be updated.\cr
Program is tested on qpdf version 8.1.0.
}
