% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_operations.R
\name{pop_pdf_info}
\alias{pop_pdf_info}
\title{Extracts metadata from pdf}
\usage{
pop_pdf_info(file)
}
\arguments{
\item{file}{complete path to the pdf file}
}
\value{
a dataframe with 5 columns:\cr
readable:  binary\cr
n_page:    integer\cr
encrypted: binary\cr
locked:    binary\cr
size:      numeric (in KB)
}
\description{
This function is part of the 'pdf operations' (pop) function set\cr
This function reads a pdf and returns metadata
}
