% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_operations.R
\name{pop_pdf_to_txt}
\alias{pop_pdf_to_txt}
\title{Converts a PDF to text file}
\usage{
pop_pdf_to_txt(file, outdir = NULL, wait = FALSE)
}
\arguments{
\item{file}{A PDF file}

\item{outdir}{The folder where the txt file should be saved\cr
(If NULL a new folder called 'txt' will be created in the same directory where the pdf
file is located)}

\item{wait}{Shall the program wait to trigger another instance of pdftotext.exe? (meaningfull if
a lot of different pdf files have to be converted, but is slower)}
}
\value{
A text file (the text file will have the same name as the pdf file)
}
\description{
This function is part of the 'pdf operations' (pop) function set\cr
This function converts pdf files into plain text files
}
