% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_term_list.R
\name{prepare_term_list}
\alias{prepare_term_list}
\title{Term List Preparation}
\usage{
prepare_term_list(terms)
}
\arguments{
\item{terms}{A string with terms to be prepsared}
}
\value{
A dataframe with standardized terms
}
\description{
This function takes a vector of terms and performs standardization and deduplication
measures on it. The output (a dataframe) contains a key (term_stand_space) which
is a deduplicated version of the terms provided and accounts for several issues
}
\examples{
prepare_term_list(c("Derecognize", "Dereconized", "Capitalize", "CAPITALIZED"))
}
