% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_csv}
\alias{read_csv}
\title{Read CSV Files}
\usage{
read_csv(path, pb = FALSE, delim = ";", col.types = "guess")
}
\arguments{
\item{path}{The path to the textfiles (get with list.files())}

\item{pb}{Shall a progress bar be shown?#'}

\item{delim}{Delimiter, (mostly ';' or ',')}

\item{col_types}{See ?readr::coltypes}
}
\value{
A dataframe with the columns of the csv file plus an additional identifier column: \cr
doc_id: The document name of the csv file
}
\description{
This function reads either a single of multiple csv files into a dataframe.
}
