% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult_csv.R
\name{read_mult_csv}
\alias{read_mult_csv}
\title{Read multiple csv files into a dataframe}
\usage{
read_mult_csv(path, recursive = FALSE, delim = ";",
  col.types = readr::cols(.default = readr::col_guess()))
}
\arguments{
\item{path}{path to the csv files}

\item{recursive}{shall subfolders be included?}

\item{delim}{Delimiter, (mostly ';' or ',')}

\item{col_types}{see readr::coltypes}
}
\value{
A dataframe
}
\description{
Read multiple csv files into a dataframe
}
