% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_words.R
\name{replace_words}
\alias{replace_words}
\title{Replacement of commonly misspelled words}
\usage{
replace_words(string, replace_table = NULL, nthreads = 1)
}
\arguments{
\item{string}{The input string}

\item{replace_table}{A datframe with two columns: \cr
1. column "word":    The word or phrase to search for in the string \cr
2. column "replace": The word or phrase which replaces the the word or phrase in
the first column \cr
Note: If no table is specified the default table will be used \cr
Note: The table will be checked for incorrect replacenemts
Note: The table will be checked for replacement dependencies (A -> B, B -> C <=> A -> C)}

\item{nthreads}{Number of threads for paralell processing (as integer) \cr
nthreads = 1: No parallel processing (default) \cr
nthreads = 0: All but one cores will be used (parallel::detectCores() - 1)}
}
\value{
A string with replaced words
}
\description{
Replacement of commonly misspelled words
}
\examples{
# Test String:
test <- c("investments", "assets")
replace_words(test)

}
