% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rop_functions.R
\name{rop_csv}
\alias{rop_csv}
\title{Read Multiple CSV files}
\usage{
rop_csv(files, doc.id = NULL, col.types = NULL, delim = ";")
}
\arguments{
\item{files}{The full paths to the files}

\item{doc.id}{Either NULL or a one-element charcter string}

\item{col.types}{Either NULL, a single character or a character string for all columns}

\item{delim}{The delimiter, default is ';'}
}
\value{
A Dataframe
}
\description{
Read Multiple CSV files
}
