% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rop_functions.R
\name{rop_rds}
\alias{rop_rds}
\title{Read Multiple RDS Files}
\usage{
rop_rds(files, doc.id = NULL)
}
\arguments{
\item{files}{The full paths to the files}

\item{doc.id}{Either NULL or a one-element charcter string}
}
\value{
A Dataframe
}
\description{
Read Multiple RDS Files
}
