% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rop_functions.R
\name{rop_txt}
\alias{rop_txt}
\title{Read Multiple Text Files}
\usage{
rop_txt(files, doc.id = NULL)
}
\arguments{
\item{files}{The full paths to the files}

\item{doc.id}{Either NULL or a one-element charcter string}
}
\value{
A Dataframe
}
\description{
Read Multiple Text Files
}
