% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_seq_det.R
\name{seq_det}
\alias{seq_det}
\title{Detect Integer Sequence in Vector}
\usage{
seq_det(vec, seq, type = c("jaap1", "jaap2", "frank"))
}
\arguments{
\item{vec}{An integer vector}

\item{seq}{A sequence of integer values to search in 'vec'}

\item{type}{Which algorithm to employ?}
}
\value{
An integer vector with the indices of the sequence
}
\description{
Detect Integer Sequence in Vector
}
\examples{
library(tpfuns)
seq_det(c(1,2,3,5,4,1), c(5,4,1), type = 1)
seq_det(c(1,2,3,5,4,1), c(5,4,1), type = 2)
seq_det(c(1,2,3,5,4,1), c(5,4,1), type = 3)
}
