% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SeqInVecOpt}
\alias{SeqInVecOpt}
\title{Find Sequence in Integer Vector}
\usage{
SeqInVecOpt(myVector, mySequence)
}
\arguments{
\item{vec}{An integer vecotr}

\item{seq}{The sequence}
}
\description{
Find Sequence in Integer Vector
}
