% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{split_chunk}
\alias{split_chunk}
\title{Wrapper Around the Split Function}
\usage{
split_chunk(x, chunk.size, type = c("sort", "equal", "rand"))
}
\arguments{
\item{x}{The object}

\item{chunk.size}{The number of elements}

\item{type}{How should the elements be ordered?}
}
\value{
A list
}
\description{
Splits an object into chunk sizes and puts them into a list
}
\examples{
libraray(tpfuns)

split_chunk(1:10, 4, "sort")
split_chunk(1:10, 4, "equal")
split_chunk(1:10, 4, "rand")
}
