% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_chars.R
\name{stand_chars}
\alias{stand_chars}
\title{Standardization of Strings}
\usage{
stand_chars(string, stand.table = NULL)
}
\arguments{
\item{string}{A string or a vector of strings to be standardized}

\item{stand.table}{A table with standardizations (needs to have a column for the
chracters that needs to be standardized (in hex values) and a column for the replacement)
See the attached table as example. If no table is specified the default table will be used}
}
\value{
A string
}
\description{
Standardization of Strings
}
