% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations.R
\name{stand_chars_fast}
\alias{stand_chars_fast}
\title{Standardization of Strings}
\usage{
stand_chars_fast(string, stand.table = NULL)
}
\arguments{
\item{string}{A string or a vector of strings to be standardized}

\item{stand.table}{A table with standardizations (needs to have a column for the
chracters that needs to be standardized (in hex values) and a column for the replacement)
See the attached table as example. If no table is specified the default table will be used}
}
\value{
A string
}
\description{
Standardization of Strings
}
\examples{
# Replacement of non-ASCII characters to its nearest ASCII equivalent
stand_chars_fast("  ĆȦpītaĻĮSȆĎ")

# Standardization of some characters
stand_chars_fast(c("A + B", "A AND B", "A & B", "( A and B  )"))
}
