% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_operations.R
\name{stand_db_names}
\alias{stand_db_names}
\title{Standardizes Names from Accounting Data Provider (e.g. Datastream)}
\usage{
stand_db_names(string, regex.table = NULL, standardized = FALSE,
  nthreads = 1)
}
\arguments{
\item{string}{String with company names to be standardized}

\item{regex.table}{A table with with regexes for string replacement, if NULL default table is used}

\item{standardized}{Is the input already standardized? If not the standardization function
stand_chars is used to standardize the input string}

\item{nthreads}{Number of threads for paralell processing (as integer) \cr
nthreads = 1: No parallel processing (default) \cr
nthreads = 0: All but one cores will be used (parallel::detectCores() - 1)}
}
\value{
A string with standardized names
}
\description{
Standardizes Names from Accounting Data Provider (e.g. Datastream)
}
\examples{
a <- "BASF SE dead 31/12/2005"
stand_db_names(a)


}
