% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_standardization.R
\name{stand_docs_fast}
\alias{stand_docs_fast}
\title{Document standardization}
\usage{
stand_docs_fast(docs, stand.table = NULL, replace.table = NULL)
}
\arguments{
\item{stand.table}{A table with standardizations (needs to have a column for the chracters that needs to
be standardized (in hex values) and a column for the replacement) \cr
See the attached table as example. If no table is specified the default table will be used}

\item{replace.table}{A table with British English and American English spelling}

\item{string}{A string}
}
\value{
A string
}
\description{
This function standardizes documents to be suitable for frequqncy counts
}
