% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_legal_entities.R
\name{stand_le_2}
\alias{stand_le_2}
\title{Standardize Legal Entities (Vectorized version)}
\usage{
stand_le_2(string, le.table = NULL, standardized = FALSE)
}
\arguments{
\item{string}{The strings containing legal entities}

\item{le.table}{A table with legal entities, if NULL the default table will be used}

\item{standardized}{Is the input standardized?}
}
\value{
A dataframe with standardized legal entities
}
\description{
Standardize Legal Entities (Vectorized version)
}
