% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_sim.R
\name{string_sim}
\alias{string_sim}
\title{String Similarity}
\usage{
string_sim(a, b, par = FALSE)
}
\arguments{
\item{a}{Source String}

\item{b}{Target String}

\item{par}{Parellel Execution?}
}
\value{
A dataframe with the highest similarity between the source and the target string
}
\description{
String Similarity
}
