% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_extraction.R
\name{tax_get_def_ifrs}
\alias{tax_get_def_ifrs}
\title{Extract Definitions from Taxonomy Files}
\usage{
tax_get_def_ifrs(doc.file)
}
\arguments{
\item{doc.file}{The xml file of the taxonomy, mostly label with doc_...}
}
\value{
A dataframe with two columns:\cr
label: The labels of the taxonomy files (used as an identifier)\cr
def:   The definitions of the taxonomy files
}
\description{
This funnction extracts definitions from taxonomy xml files.\cr
(Tested on US GAAP, UK GAAP and IFRS taxonomies)
}
