% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_extraction.R
\name{tax_get_ref_us_gaap}
\alias{tax_get_ref_us_gaap}
\title{Extract References from Taxonomy Files}
\usage{
tax_get_ref_us_gaap(ref.file)
}
\arguments{
\item{ref.file}{The xml file of the taxonomy, mostly label with ref_...}
}
\value{
A dataframe with several reference columns and one identification column used for
reconsilation to the terms and definition files from the functions tax_get_term()
and tax_get_def()
}
\description{
Extract References from Taxonomy Files
}
