% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_extraction.R
\name{tax_get_term_def}
\alias{tax_get_term_def}
\title{Extract Terms and Definitions from Taxonomy Files}
\usage{
tax_get_term_def(lab.file, doc.file)
}
\arguments{
\item{lab.file}{The xml file of the taxonomy, mostly label with lab_...}

\item{doc.file}{The xml file of the taxonomy, mostly label with doc_...}
}
\value{
A dataframe with four columns:\cr
label: The labels of the taxonomy files (used as an identifier)\cr
term:  The terms of the taxonomy files
role:  An identification of the type of the term
def:   The definitions of the taxonomy files
}
\description{
This funnction conbines the tax_get_term() and tax_get_def() functions\cr
}
