% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_extraction.R
\name{tax_get_term_ifrs}
\alias{tax_get_term_ifrs}
\title{Extract Terms from Taxonomy Files}
\usage{
tax_get_term_ifrs(lab.file, pb = FALSE)
}
\arguments{
\item{lab}{The xml file of the taxonomy, mostly label with lab_...}
}
\value{
A dataframe with three columns:\cr
label: The labels of the taxonomy files (used as an identifier)\cr
term:  The terms of the taxonomy files
role:  An identification of the type of the term
}
\description{
This funnction extracts terms from taxonomy xml files.\cr
(Tested on US GAAP, UK GAAP and IFRS taxonomies)
}
