% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_extraction.R
\name{tax_levels}
\alias{tax_levels}
\title{Level Determination of Strings - Substrings}
\usage{
tax_levels(string)
}
\arguments{
\item{string}{A character string containing terms}
}
\value{
A dataframe with two columns:\cr
term: the terms of the original string
level: the level assignement
}
\description{
Level Determination of Strings - Substrings
}
