% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations.R
\name{term_operation_fast}
\alias{term_operation_fast}
\title{Perform multiple operations on terms}
\usage{
term_operation_fast(string, stand.table = NULL, replace.table = NULL)
}
\arguments{
\item{string}{A string}

\item{stand.table}{A table with standardizations (needs to have a column for the chracters that needs to
be standardized (in hex values) and a column for the replacement) \cr
See the attached table as example. If no table is specified the default table will be used}

\item{replace.table}{A table with British English and American English spelling}
}
\value{
A string
}
\description{
This function combines the stand_chars_fast, americanize_fast, and lemmatize_fast function
}
\examples{
term_operation_fast(c("  ĆȦpītaĻĮSȆĎ", "centres"))
}
