% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations_2.R
\name{tok_pos_lem}
\alias{tok_pos_lem}
\title{Helper Function for top_lem_doc() and top_lem}
\usage{
tok_pos_lem(string)
}
\arguments{
\item{string}{A string}
}
\value{
A dataframe
}
\description{
Helper Function for top_lem_doc() and top_lem
}
