% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations_2.R
\name{top_americanize}
\alias{top_americanize}
\title{Americanize Words or Strings}
\usage{
top_americanize(string)
}
\arguments{
\item{string}{A charcter string}
}
\value{
A character string with only American English spelling
}
\description{
This function is part of the 'term operations' (top) function set\cr
This function converts words in character strings from British English spelling to
American English Spelling
}
\examples{
americanize(c("summarised, assets", "yoghurt"))
}
