% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations.R
\name{top_lem_doc}
\alias{top_lem_doc}
\title{Tokenize and Lemmatize Documents (with sentence split)}
\usage{
top_lem_doc(string)
}
\arguments{
\item{string}{The input document as a charcter string}
}
\value{
A list with tokenized words and sentence split
}
\description{
Tokenize and Lemmatize Documents (with sentence split)
}
