% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations.R
\name{top_lem_term}
\alias{top_lem_term}
\title{Tokenize and Lemmatize Terms}
\usage{
top_lem_term(string)
}
\arguments{
\item{string}{A string of terms}
}
\value{
A list with terms and tokens
}
\description{
Tokenize and Lemmatize Terms
}
