% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_functions.R
\name{top_lookup_wrap}
\alias{top_lookup_wrap}
\title{Wrapper around the top_term_lookup function}
\usage{
top_lookup_wrap(string, type = c("lemmatize", "americanize", "errors",
  "split", "all"), tokenize = FALSE)
}
\arguments{
\item{string}{Input string}

\item{type}{Determines which lookup table to use}

\item{tokenize}{If TRUE terms will be first split by word boundaries and than put back together}
}
\value{
A new vector with reassigned values
}
\description{
Wrapper around the top_term_lookup function
}
