% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_functions.R
\name{top_punct}
\alias{top_punct}
\title{Punctuation Operations}
\usage{
top_punct(string, lower = TRUE, punct.stand = TRUE,
  punct.rem = c("none", "partial", "complete"), punct.sub = FALSE)
}
\arguments{
\item{string}{A character string}

\item{lower}{Shall the string be transfored to lower case? (TRUE/FALSE) \cr
This option is only available if punct.stand == TRUE}

\item{punct.stand}{Shall the input string be standardized? (TRUE/FALSE)}

\item{punct.rem}{Shall punctuation be removed? (One of none/partial/complete) \cr
none:     No removal of punctuations\cr
partial:  All punctuation except ampersand (&), forward slash (/) and hyphen (-) is removed\cr
complete: All punctuation is removed}

\item{punct.sub}{Shall certein punctuation characters be replaced? (TRUE/FALSE) \cr
If TRUE the ampersand (&) and plus (+) characters will be replace with 'and'}
}
\value{
A charcter string
}
\description{
This function is part of the 'term operations' (top) function set\cr
It takes and input string and standardizes / substitutes / replaces punctuations and
characters
}
\examples{
library(tpfuns)
input.string <- "TeSt STring: ( This, ) String [ Is a ] Test & ít-is badlý written"

top_punct(input.string, TRUE, TRUE, punct.rem = "none", FALSE)
top_punct(input.string, TRUE, TRUE, "partial", FALSE)
top_punct(input.string, TRUE, TRUE, "complete", FALSE)

top_punct(input.string, TRUE, TRUE, "partial", TRUE)
}
