% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations_2.R
\name{top_rem_punct}
\alias{top_rem_punct}
\title{Removes Punctuation in Words or Strings}
\usage{
top_rem_punct(string, punct.replacement = FALSE)
}
\arguments{
\item{string}{A charcter string}

\item{punct.replacement}{Shall specific punctuation be replaced?\cr
currently implemented: & -> and; + -> and}
}
\value{
A character string without punctuations
}
\description{
This function is part of the 'term operations' (top) function set\cr
This function takes a charcter string and removes all punctuations (except the single
quote " ' " which is completely removed, all other punctuations are replaced by a space
character)
}
\examples{
top_rem_punct("David's assets,, are   boring")
}
