% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations.R
\name{top_repl_strings}
\alias{top_repl_strings}
\title{Replace strings}
\usage{
top_repl_strings(table, col, string.match, string.repl)
}
\arguments{
\item{table}{a}

\item{col}{a}

\item{string.match}{a}

\item{string.repl}{a}
}
\value{
a
}
\description{
Replace strings
}
\examples{
a
}
