% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_operations.R
\name{top_stand_punct}
\alias{top_stand_punct}
\title{Standardize Punctuation for Words or Strings}
\usage{
top_stand_punct(string, lower = TRUE, ascii = TRUE)
}
\arguments{
\item{string}{A charcter string}

\item{lower}{Shall the string be lowered?}

\item{ascii}{Shall non-ascii characters be converted to ascii characters?}
}
\value{
A standardized character string
}
\description{
This function is part of the 'term operations' (top) function set\cr
This function takes a charcter string and standardizes punctuation. (e.g. it ensures
correct placements of hyphens, standardizes space characters, lowers and trims strings,
...)
}
\examples{
top_stand_punct("Hallo    Paul , I´´m cUrrently looking for money ( cash   ) [ or equivalents ] ")
}
