% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_functions.R
\name{top_term_lookup}
\alias{top_term_lookup}
\title{Hash Table/Dictionary Lookup}
\usage{
top_term_lookup(terms, key.match, key.reassign = NULL,
  tokenize = FALSE)
}
\arguments{
\item{terms}{A vector of terms to undergo a lookup}

\item{key.match}{Takes one of the following:\cr
(1) a two column data.frame of a match key and reassignment column\cr
(2) a named list of vectors (Note: if data.frame or named list supplied no key reassign needed)\cr
(3) a single vector match key.}

\item{key.reassign}{A single reassignment vector supplied if key.match is not a two column data.frame/named list}

\item{tokenize}{If TRUE terms will be first split by word boundaries and than put back together}
}
\value{
A new vector with reassigned values
}
\description{
Replaces the individual elements of the vector 'terms' by the reassign key
(largely based on qdapTools::lookup)
}
\examples{
library(tpfuns)
a <- c("property and plant", "property", "plant")
b <- c("property", "plant")
c <- c("replace 1", "replace 2")
top_term_lookup(a, b, c)
top_term_lookup(a, b, c, TRUE)
}
