% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox_functions.R
\name{tox_doc_stats}
\alias{tox_doc_stats}
\title{Text Statistics for Documents}
\usage{
tox_doc_stats(text, lan = c("single", "mixed"))
}
\arguments{
\item{text}{a character string of texts}

\item{lan}{If the language of the text is unknown it can be determined}
}
\value{
A Dataframe with the following columns:
doc_id: the filename of the text file\cr
n_page: number of pages\cr
n_word: number of words\cr
n_char: number of character\cr
r_word_page: ratio of words per page\cr
r_char_word: ratio of characters per word\cr
size: size of the text file in KB\cr
lan_1: first detected language (optional)\cr
lan_2: second detected language (optional)\cr
rel_1: reliability of the first detected language (optional)\cr
rel_2: reliability of the second detected language (optional)
}
\description{
This function is part of the 'text operation' (tox) dunction set\cr
This function counts pages, words and characters, it determines the filesize and
calculates ratios. (optional the language of the text is returned)
}
