% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{unique_pairs}
\alias{unique_pairs}
\title{Unique Combinations of all Elements within a Vector}
\usage{
unique_pairs(vec)
}
\arguments{
\item{vec}{}
}
\value{
A Dataframe with (n^2-n)/2 observations
}
\description{
Given a vector of elements this function calculates the unique combinations between all
element
}
\examples{
# unique pairs of 3 companies, a dataframe with (3^2-3)/2 = 3 observations is returned
unique_pairs(c("comp_1", "comp_2", "comp_3"))
}
