% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{write_csv_rds}
\alias{write_csv_rds}
\title{Function that saves files both as csv and rds}
\usage{
write_csv_rds(x, path, format = c("csv", "rds"))
}
\arguments{
\item{x}{A dataframe}

\item{path}{The full path to the file (without file endings)}
}
\value{
two saved files
}
\description{
Function that saves files both as csv and rds
}
