% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_operations.R
\name{cop_match_comp}
\alias{cop_match_comp}
\title{Matches Companies against Datastream and Orbis}
\usage{
cop_match_comp(names, match.table = NULL, cols.ident = NULL,
  match.type = c("full", "sub", "approx"))
}
\arguments{
\item{match.table}{The match table (not part of the package, needs to be defined seperately!)}

\item{match.type}{Type of match, any combination of 'full', 'sub' and 'approx'}

\item{string}{A character string with company names to match}
}
\value{
A dtaframe with matched companies
}
\description{
Matches Companies against Datastream and Orbis
}
