% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_match.R
\name{copm_prep_names}
\alias{copm_prep_names}
\title{Name Preparation for Company Matching Procedure}
\usage{
copm_prep_names(names, adj.regex = "default")
}
\arguments{
\item{names}{A character string of company names}

\item{adj.regex}{Adjustment Regex}
}
\value{
A dataframe with different 3 columns containing different name varieties.\cr
Col 1 ('n0'):    Original Name of the company.\cr
Col 2 ('n1'):    Adjusted Name of the company.\cr
Col 3 ('ntype'): Adjustemt type.
}
\description{
This function takes a character string and prepares a deduplicated list with maximum 24
different variations of the name. The different name varieties are stored in the 'ntype'
column:\cr
1:	 original name (standardized punctuation) +  standardized words + americanization\cr
2:	 original name (standardized punctuation) +  standardized words + americanization + standardized legal entities\cr
3:	 original name (standardized punctuation) +  standardized words + americanization + removed legal entities\cr
4:	 original name (standardized punctuation) +  standardized words + americanization + removed brackets\cr
5:	 original name (standardized punctuation) +  standardized words + americanization + removed brackets + standardized legal entities\cr
6:	 original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed legal entities\cr
7:	 original name (standardized punctuation) +  standardized words + americanization + removed terms\cr
8:	 original name (standardized punctuation) +  standardized words + americanization + removed terms + standardized legal entities\cr
9:	 original name (standardized punctuation) +  standardized words + americanization + removed terms + removed legal entities\cr
10: original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed terms\cr
11: original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed terms + standardized legal entities\cr
12: original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed terms + removed legal entities\cr
13: original name (standardized punctuation) +  standardized words + americanization + removed space characters\cr
14: original name (standardized punctuation) +  standardized words + americanization + standardized legal entities + removed space characters\cr
15: original name (standardized punctuation) +  standardized words + americanization + removed legal entities + removed space characters\cr
16: original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed space characters\cr
17: original name (standardized punctuation) +  standardized words + americanization + removed brackets + standardized legal entities + removed space characters\cr
18: original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed legal entities + removed space characters\cr
19: original name (standardized punctuation) +  standardized words + americanization + removed terms + removed space characters\cr
20: original name (standardized punctuation) +  standardized words + americanization + removed terms + standardized legal entities + removed space characters\cr
21: original name (standardized punctuation) +  standardized words + americanization + removed terms + removed legal entities + removed space characters\cr
22: original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed terms + removed space characters\cr
23: original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed terms + standardized legal entities + removed space characters\cr
24: original name (standardized punctuation) +  standardized words + americanization + removed brackets + removed terms + removed legal entities + removed space characters\cr
}
