% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{cos_string_pair}
\alias{cos_string_pair}
\title{COsine Similarity between two Strings}
\usage{
cos_string_pair(a, b)
}
\arguments{
\item{a}{A String (or a vector of strings)}

\item{b}{A String (or a vector of strings)}
}
\value{
The cosine similarity (either a single number or a vector of numbers)
}
\description{
COsine Similarity between two Strings
}
