% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mes_functions.R
\name{mes_pbs}
\alias{mes_pbs}
\title{Profile Based Standardization}
\usage{
mes_pbs(freq.table, return = c("normal", "matrix"))
}
\arguments{
\item{freq.table}{A Table}

\item{return}{Table or Matrix}
}
\value{
Table or Matrix
}
\description{
Profile Based Standardization
}
\examples{
library(tpfuns)
freq.table <- tibble::tibble(
doc_id = c(1,1,1,1,1,2,2,2,2,3,3,3,3),
con_id = c(1,1,2,2,2,1,1,3,3,1,3,2,3),
term   = c("a", "b", "c", "d", "e", "a", "b", "f", "g", "a", "f", "e", "g"),
freq   = c(21,12,58,32,14,21,14,66,14,12,85,100,12))

mes_pbs(freq.table)
}
