% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_txt}
\alias{read_txt}
\title{Read Text Files}
\usage{
read_txt(path, pb = FALSE)
}
\arguments{
\item{path}{The path to the textfiles (get with list.files())}

\item{pb}{Shall a progress bar be shown?}
}
\value{
A dataframe with two columns: \cr
doc_id: The document name of the text file
text:   The text of the text file
}
\description{
This function reads either a single of multiple text files into a dataframe. The content
of each textfile is stored in a single string in one row of the datframe
}
