% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_db_names.R
\name{stand_db_names}
\alias{stand_db_names}
\title{Standardizes Names from Accounting Data Provider (e.g. Datastream)}
\usage{
stand_db_names(string, regex.table = NULL, standardized = FALSE)
}
\arguments{
\item{string}{String with company names to be standardized}
}
\value{
A string with standardized names
}
\description{
Standardizes Names from Accounting Data Provider (e.g. Datastream)
}
