% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_legal_entities.R
\name{stand_le_2}
\alias{stand_le_2}
\title{Standardize Legal Entities (Vectorized version)}
\usage{
stand_le_2(string, le.table = NULL, standardized = FALSE,
  nthreads = 1)
}
\arguments{
\item{string}{The strings containing legal entities}

\item{le.table}{A table with legal entities, if NULL the default table will be used}

\item{standardized}{Is the input already standardized? If not the standardization function
stand_chars is used to standardize the input string}

\item{nthreads}{Number of threads for paralell processing (as integer) \cr
nthreads = 1: No parallel processing (default) \cr
nthreads = 0: All but one cores will be used (parallel::detectCores() - 1)}
}
\value{
A string with legal entities removed
}
\description{
Standardize Legal Entities (Vectorized version)
}
\examples{
a <- c("BASF SE", "BASF AG")
stand_le_2(a)

}
